Add connection timeout to connect_with_connector_lazy (#1619)
Currently connect_with_connector_lazy doesn't respect connection
timeouts. This means that if you have a misbehaving server a request
that's connecting lazily can hang for the client without timing out.
This commit adds the timeout to the custom connector if it has been set
on the endpoint, bringing the behaviour in line with other connect
methods.