transport: extend ClientTlsConfig to omit ALPN (#1640)
Sometimes servers are secured with tls and speak gRPC, but don't perform
ALPN protocol negotation. Most other gRPC implementations out there, as
far as I can tell will just assume that the remote server is speaking h2
if there is ALPN. Tonic is strict in this regard. This patch takes the
conservative approach of allowing users to opt into assuming that the
remote server is running h2. A more aggressive patch in the future might
be to invert the default.