make k8s edits to enable replica load balancing
The only setting that matters should be `clusterIP: None`. This sets a service to be "headless."

With a headless service, when inquiring k8s's DNS, it returns a list of IPs corresponding to service replicas. This enables client-side LB.

`sessionAffinity: None` was expected to work. However, even when creating multiple connections with our Tonic client, it doesn't actually connect to multiple replicas.